added memory

+++ source/main.cpp
@@ -5,13 +5,13 @@ int main(int argc, char** argv) {
 
 	FibNumbers fibNum;
 
-	int n = 0;
+	long int n = 0;
 
 	std::cout << "Введите число: ";
 	std::cin >> n;
 
 	fibNum.getQuantityFibNumbers(n);
 
-
+	
 	return 0;
 }
